Control commit for experts
Many changes, in two words - continue to work with task
See also: https://docs.google.com/document/d/1cUVHZfEo4uMGJBBADbRMQzdTjSsySiFJ3xVvZ8DsQ7o/edit?tab=t.0

+++ src/main/java/org/example/animalchipization/service/animal/impl/AnimalServiceImpl.java
@@ -1,13 +1,16 @@
 package org.example.animalchipization.service.animal.impl;
 
-import org.example.animalchipization.dto.animal.AnimalDtoIn;
-import org.example.animalchipization.dto.animal.AnimalDtoOut;
-import org.example.animalchipization.dto.animal.AnimalDtoUpdate;
-import org.example.animalchipization.dto.animal.AnimalSearchCriteria;
+import jakarta.persistence.criteria.JoinType;
+import org.example.animalchipization.dto.animal.*;
+import org.example.animalchipization.dto.location.LocationDtoOut;
+import org.example.animalchipization.dto.location.LocationSearchCriteria;
 import org.example.animalchipization.entities.Animal;
 import org.example.animalchipization.entities.AnimalType;
+import org.example.animalchipization.entities.Location;
+import org.example.animalchipization.enums.AnimalLifeStatus;
 import org.example.animalchipization.enums.errors.AnimalError;
 import org.example.animalchipization.exception.entities.AnimalException;
+import org.example.animalchipization.mappers.LocationMapper;
 import org.example.animalchipization.mappers.animal.AnimalMapper;
 import org.example.animalchipization.repository.AccountRepository;
 import org.example.animalchipization.repository.AnimalRepository;
@@ -36,17 +39,17 @@ public class AnimalServiceImpl implements AnimalService {
     private final LocationRepository locationRepository;
     private final AnimalTypeRepository animalTypeRepository;
     private final AnimalMapper animalMapper;
-    private final JpaSpecificationBuilder<Animal> jpaSpecificationBuilder;
+    private final LocationMapper locationMapper;
 
 
     @Autowired
-    public AnimalServiceImpl(AnimalRepository animalRepository, AccountRepository accountRepository, AnimalTypeRepository animalTypeRepository, AnimalMapper animalMapper, JpaSpecificationBuilder<Animal> jpaSpecificationBuilder, LocationRepository locationRepository) {
+    public AnimalServiceImpl(AnimalRepository animalRepository, AccountRepository accountRepository, AnimalTypeRepository animalTypeRepository, AnimalMapper animalMapper, LocationRepository locationRepository, LocationMapper locationMapper) {
         this.animalRepository = animalRepository;
         this.accountRepository = accountRepository;
         this.animalTypeRepository = animalTypeRepository;
         this.animalMapper = animalMapper;
-        this.jpaSpecificationBuilder = jpaSpecificationBuilder;
         this.locationRepository = locationRepository;
+        this.locationMapper = locationMapper;
     }
 
 
@@ -67,7 +70,7 @@ public class AnimalServiceImpl implements AnimalService {
         Set<AnimalType> animalTypes = new HashSet<>(
                 animalTypeRepository.findAllById(animalDtoIn.getAnimalTypes())
         );
-        if (animalTypes.size() < animalDtoIn.getAnimalTypes().size()){
+        if (animalTypes.size() < animalDtoIn.getAnimalTypes().size()) {
             throw new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND);
         }
         if (!accountRepository.existsAccountByAccountId(animalDtoIn.getChipperId())) {
@@ -92,6 +95,22 @@ public class AnimalServiceImpl implements AnimalService {
         Animal existingAnimal = animalRepository.findById(animalId)
                 .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_NOT_FOUND));
 
+        if (existingAnimal.getLifeStatus() == AnimalLifeStatus.DEAD &&
+                animalDtoUpdate.getLifeStatus() == AnimalLifeStatus.ALIVE) {
+            throw new AnimalException(AnimalError.ANIMAL_ALREADY_DEAD);
+        }
+        if (existingAnimal.getChippingLocationId().getLocationId()
+                .equals(animalDtoUpdate.getChippingLocationId())) {
+            throw new AnimalException(AnimalError.ANIMAL_CHIPPING_LOCATION_ALREADY_EXISTS);
+        }
+
+        if (!accountRepository.existsAccountByAccountId(animalDtoUpdate.getChipperId())) {
+            throw new AnimalException(AnimalError.ANIMAL_CHIPPER_NOT_FOUND);
+        }
+        if (!locationRepository.existsLocationByLocationId(animalDtoUpdate.getChippingLocationId())) {
+            throw new AnimalException(AnimalError.ANIMAL_CHIPPING_LOCATION_NOT_FOUND);
+        }
+
         animalMapper.updateEntity(existingAnimal, animalDtoUpdate);
 
         Animal savedAnimal = animalRepository.save(existingAnimal);
@@ -102,35 +121,40 @@ public class AnimalServiceImpl implements AnimalService {
 
     @Override
     public void deleteAnimalById(Long animalId) {
+
+        if (!animalRepository.existsById(animalId)) {
+            throw new AnimalException(AnimalError.ANIMAL_NOT_FOUND);
+        }
+
         animalRepository.deleteById(animalId);
     }
 
     @Override
     @Transactional
-    public List<AnimalDtoOut> searchAnimal(AnimalSearchCriteria animalSearchCriteria, Pageable pageable) {
+    public List<AnimalDtoOut> searchAnimals(AnimalSearchCriteria animalSearchCriteria, Pageable pageable) {
 
         Specification<Animal> spec = Specification.where(
-                jpaSpecificationBuilder.dateTimeFrom(
+                JpaSpecificationBuilder.<Animal>dateTimeFrom(
                         "chippingDateTime",
                         animalSearchCriteria.startDateTime())
         ).and(
-                jpaSpecificationBuilder.dateTimeTo(
+                JpaSpecificationBuilder.dateTimeTo(
                         "chippingDateTime",
                         animalSearchCriteria.endDateTime())
         ).and(
-                jpaSpecificationBuilder.equal(
+                JpaSpecificationBuilder.equal(
                         "chipperId",
                         animalSearchCriteria.chipperId())
         ).and(
-                jpaSpecificationBuilder.equal(
+                JpaSpecificationBuilder.equal(
                         "chippingLocationId",
                         animalSearchCriteria.chippingLocationId())
         ).and(
-                jpaSpecificationBuilder.equal(
+                JpaSpecificationBuilder.equal(
                         "lifeStatus",
                         animalSearchCriteria.lifeStatus())
         ).and(
-                jpaSpecificationBuilder.equal(
+                JpaSpecificationBuilder.equal(
                         "gender",
                         animalSearchCriteria.gender())
         );
@@ -138,5 +162,110 @@ public class AnimalServiceImpl implements AnimalService {
 
 
         return animalPage.map(animalMapper::toDto).getContent();
+    }
+
+    @Override
+    @Transactional
+    public List<LocationDtoOut> searchLocations(LocationSearchCriteria locationSearchCriteria, Pageable pageable) {
+
+        Specification<Location> spec = Specification.where(
+                JpaSpecificationBuilder.<Location>dateTimeFrom(
+                        "chippingDateTime",
+                        locationSearchCriteria.startDateTime())
+        ).and(
+                JpaSpecificationBuilder.dateTimeTo(
+                        "chippingDateTime",
+                        locationSearchCriteria.endDateTime())
+        ).and(
+                JpaSpecificationBuilder.joinManyToMany(
+                        "animals",
+                        "animalId", locationSearchCriteria.animalId(), JoinType.INNER)
+        );
+        Page<Location> locationPage = locationRepository.findAll(spec, pageable);
+
+        return locationPage.map(locationMapper::toDto).getContent();
+    }
+
+    @Override
+    @Transactional
+    public AnimalDtoOut addAnimalType(Long animalId, Long animalTypeId) {
+
+        Animal existingAnimal = animalRepository.findById(animalId)
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_NOT_FOUND));
+
+        AnimalType existingType = animalTypeRepository.findById(animalTypeId)
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND));
+
+        Set<AnimalType> existingAnimalTypes = existingAnimal.getAnimalTypes();
+
+        if (existingAnimalTypes.contains(existingType)) {
+            throw new AnimalException(AnimalError.ANIMAL_TYPES_DUPLICATES);
+        }
+
+        existingAnimalTypes.add(existingType);
+
+        existingAnimal.setAnimalTypes(existingAnimalTypes);
+
+        Animal savedAnimal = animalRepository.save(existingAnimal);
+
+        return animalMapper.toDto(savedAnimal);
+
+    }
+
+    @Override
+    @Transactional
+    public AnimalDtoOut updateAnimalType(Long animalId, UpdateAnimalTypeDto updateAnimalTypeDto) {
+
+        Animal existingAnimal = animalRepository.findById(animalId)
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_NOT_FOUND));
+
+        AnimalType existingType = animalTypeRepository.findById(updateAnimalTypeDto.getOldTypeId())
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND));
+
+        AnimalType newType = animalTypeRepository.findById(updateAnimalTypeDto.getNewTypeId())
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND));
+
+        Set<AnimalType> existingAnimalTypes = existingAnimal.getAnimalTypes();
+
+        if (!existingAnimalTypes.contains(existingType)) {
+            throw new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND);
+        }
+        if (existingAnimalTypes.contains(newType)) {
+            throw new AnimalException(AnimalError.ANIMAL_TYPES_DUPLICATES);
+        }
+
+        existingAnimalTypes.remove(existingType);
+        existingAnimalTypes.add(newType);
+
+        existingAnimal.setAnimalTypes(existingAnimalTypes);
+
+        Animal savedAnimal = animalRepository.save(existingAnimal);
+
+        return animalMapper.toDto(savedAnimal);
+    }
+
+    @Override
+    @Transactional
+    public AnimalDtoOut deleteAnimalType(Long animalId, Long animalTypeId) {
+
+        Animal existingAnimal = animalRepository.findById(animalId)
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_NOT_FOUND));
+
+        AnimalType existingType = animalTypeRepository.findById(animalTypeId)
+                .orElseThrow(() -> new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND));
+
+        Set<AnimalType> existingAnimalTypes = existingAnimal.getAnimalTypes();
+
+        if (!existingAnimalTypes.contains(existingType)) {
+            throw new AnimalException(AnimalError.ANIMAL_TYPE_NOT_FOUND);
+        }
+
+        existingAnimalTypes.remove(existingType);
+
+        existingAnimal.setAnimalTypes(existingAnimalTypes);
+
+        Animal savedAnimal = animalRepository.save(existingAnimal);
+
+        return animalMapper.toDto(savedAnimal);
     }
 }
